fix: Add api controller and fix fetch api in email-verification
