+++ src/main/java/com/doci/webPrj/config/SpringSecurityConfig.java
@@ -25,7 +25,7 @@ public class SpringSecurityConfig {
         http.csrf(csrf -> csrf.disable()).cors(cors -> cors.disable())
                 .authorizeHttpRequests(request -> request
                         .dispatcherTypeMatchers(DispatcherType.FORWARD).permitAll()
-                        .requestMatchers("/image/**", "/css/**", "/js/**","/join/**").permitAll()
+                        .requestMatchers("/image/**", "/css/**", "/js/**","/join/**","api/join/**").permitAll()
                         .requestMatchers("/admin/**").hasRole("ADMIN")
                         .requestMatchers("/**").hasRole("USER")
                         .anyRequest().authenticated()
